convert portfolio to ssr
