+++ app/components/nfts/NFTCard.tsx
