+++ app/components/nfts/NFTTable.tsx
